addres review

+++ src/value.c
@@ -533,7 +533,7 @@ int SIValue_Compare(const SIValue a, const SIValue b, int *disjointOrNull) {
 			break;
 		case T_POINT:
 		{
-			float lon_diff = SAFE_COMPARISON_RESULT(Point_lon(a) - Point_lon(b));
+			int lon_diff = SAFE_COMPARISON_RESULT(Point_lon(a) - Point_lon(b));
 			if(lon_diff == 0)
 				return SAFE_COMPARISON_RESULT(Point_lat(a) - Point_lat(b));
 			return lon_diff;
